Update README.md
The SMS Application is a project developed during the 6th semester of my college program, showcasing my skills in Java, XML, and Android XML files. The main objective of this application is to provide users with a user-friendly interface to send and receive SMS messages on their Android devices. The app allows users to efficiently manage their SMS communications and enhances their messaging experience.